Dan McGee <dpmcgee@gmail.com> * Removed some unnecessary headers and library links * Made things static if possible * Cleaned up makefiles a bit * Fixed some old comments in the code * Fixed some errors the static code checker splint pointed out * Backwards arguments in a memset call in _alpm_db_read (could have been worse) * Other various small fixes
Other:
* Default to 80 columns when getcols cannot determine display width
* Removal of ._install as a valid install file in packages

+++ src/pacman/package.c
@@ -218,7 +218,7 @@ void dump_pkg_changelog(char *clfile, const char *pkgname)
 	{
 		while(!feof(fp))
 		{
-			fgets(line, PATH_MAX, fp);
+			fgets(line, (int)PATH_MAX, fp);
 			printf("%s", line);
 			line[0] = '\0';
 		}
